@@ -12,11 +12,11 @@
     "bootstrap": "^4.4.1",
     "graphql": "^16.3.0",
     "jwt-decode": "^2.2.0",
-    "react": "^16.13.1",
+    "react": "^17.0.2",
     "react-bootstrap": "^1.0.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
