@@ -1,21 +1,53 @@
 import React from 'react';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import { ApolloClient, ApolloProvider, createHttpLink, InMemoryCache } from '@apollo/client';
+import { setContext } from "@apollo/client/link/context";
 import SearchBooks from './pages/SearchBooks';
 import SavedBooks from './pages/SavedBooks';
 import Navbar from './components/Navbar';
 
+
+//=====================//
+//      Functions      //
+//=====================//
+
+// Creates Apollo Client http link for GraphQL operations
+const httpLink = createHttpLink({
+  uri: "/graphql"
+});
+
+// Sets authentication into context
+const authLink = setContext((_, { headers }) => {
+  const token = localStorage.getItem("id_token");
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : ""
+    }
+  }
+});
+
+// Instantiates the client object and the cache object with some specific options
+const client = new ApolloClient({
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache({})
+});
+
+
 function App() {
   return (
-    <Router>
-      <>
-        <Navbar />
-        <Switch>
-          <Route exact path='/' component={SearchBooks} />
-          <Route exact path='/saved' component={SavedBooks} />
-          <Route render={() => <h1 className='display-2'>Wrong page!</h1>} />
-        </Switch>
-      </>
-    </Router>
+    <ApolloProvider client={client}>
+      <Router>
+        <>
+          <Navbar />
+          <Routes>
+            <Route exact path='/' element={<SearchBooks />} />
+            <Route exact path='/saved' element={<SavedBooks />} />
+            <Route render={() => <h1 className='display-2'>Wrong page!</h1>} />
+          </Routes>
+        </>
+      </Router>
+    </ApolloProvider>
   );
 }
 
